Replace deprecated media library, upgrade Rust/Dart libs

+++ linux/packaging/appimage/bundle_deps.ps1
@@ -1,7 +1,7 @@
 param ([string] $executablePath, [string] $copyTo)
 
-$libc_libs = "ld-linux-x86-64", "libBrokenLocale", "libanl", "libc", "libc_malloc_debug", "libdl", "libm", "libmemusage", "libmvec", "libnsl", "libnss_compat", "libnss_dns", "libnss_files", "libnss_hesiod", "libpcprofile", "libpthread", "libresolv", "librt", "libthread_db", "libutil"
-$skip_libs = $libc_libs # + $other_libs + $more_other_libs
+$libc_libs = "ld-linux-x86-64", "libBrokenLocale", "libanl", "libc", "libc_malloc_debug", "libdl", "libm", "libmemusage", "libmvec", "libnsl", "libnss_compat", "libnss_dns", "libnss_files", "libnss_hesiod", "libpcprofile", "libpthread", "libresolv", "librt", "libthread_db", "libutil", "libstdc++"
+$skip_libs = $libc_libs
 
 $absolute_path = Resolve-Path $executablePath
 $absolute_copy_to_path = Resolve-Path $copyTo
@@ -24,19 +24,29 @@ $lddtree_output = lddtree "$absolute_path"
 $lddtree_output_split = $lddtree_output.Split([Environment]::NewLine)
 
 $libs_to_copy = @()
+$libs_not_found = @()
 Set-Location $absolute_copy_to_path
 foreach ($lddtree_line in $lddtree_output_split) {
     $lddtree_line_split = $lddtree_line.Split("=>")
+
+    if ($lddtree_line_split[1].Trim() -eq 'not found') {
+        $lib_name = $lddtree_line_split[0].Trim()
+        if (!$libs_not_found.Contains($lib_name)) {
+            $libs_not_found += $lib_name
+        }
+        continue
+    }
+
     $lib_path = $lddtree_line_split[1].Trim().Replace(" (DEPENDENCY CYCLE)", "")
 
     if (!$libs_to_copy.contains($lib_path)) {
         $lib_name = (Get-Item $lib_path).Name.ToString().Split(".")[0]
 
         if ($lib_path.Contains($absolute_path_directory)) {
-            Write-Output "Skipping (found at destination): $lib_path"
+            # Write-Output "Skipping (found at destination): $lib_path"
             continue
         } elseif ($skip_libs.Contains($lib_name)) {
-            Write-Output "Skipping (blocklist): $lib_path"
+            # Write-Output "Skipping (blocklist): $lib_path"
             continue
         }
 
@@ -48,11 +58,21 @@ foreach ($lddtree_line in $lddtree_output_split) {
         $symlink_src = $lddtree_line_split[0].Trim()
         $symlink_dst = (Get-Item $lib_path).Name.ToString()
         if ($symlink_src -ne $symlink_dst) {
-            Write-Output "Will symlink $symlink_src to $symlink_dst"
+            if (Test-Path $symlink_src) {
+                Write-Output "Symlink already exists"
+                continue;
+            }
+            Write-Output "Symlinking $symlink_src to $symlink_dst"
             ln -s -r "$symlink_dst" "$symlink_src"
         }
     }
 }
 
-Write-Output "Bundling done!"
+if ($libs_not_found.count -gt 0) {
+    Write-Output ""
+    Write-Output "The following libs could not be found:"
+    Write-Output ($libs_not_found | Sort-Object)
+}
+
 Write-Output ""
+Write-Output "Bundling done!"
